requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
flask-jwt-extended==4.4.4
PyMySQL==1.0.3
bcrypt==4.0.1
python-dotenv==1.0.0
